Testing Slack integration while GitHub is up
When I previously tested the integration with Slack (sending reports of changes to the code of conduct here to the DH Slack #codeofconduct channel), it didn't work, but I also heard GitHub had an outage that day. So, trying this again while GitHub is up before tinkering with other settings. This commit adds a single blank space to the code of conduct.