Additional information in "Consequences" section
Now reads: "Participants asked to stop any harassing behavior are expected to comply immediately. Messages deleted by a moderator should not be reposted.
If a participant engages in harassing behavior, moderators may take any action they deem appropriate. Moderator responses may include deleting a Slack message (including replacing the message with information on why such messages are not allowed), DMing the participant to inform them of their code of conduct violation and requiring acknowledgement that a second violation will result in an immediate and permanent ban, immediate and permanent expulsion from the DH Slack, and identification of the participant as a harasser to other moderators, the DH Slack community, or the general public. "