Added intro summary & small text changes
Copied some of the foundational text from later in the document and placed it also at the top of the document (it's a long document and growing; I want to make sure the part mostly likely to be read covers the highlights). I also changed "discretion of the team owner" from owner to moderators, and added the new clause at the end of this sentence: "Harassment and other code of conduct violations reduce the value of our community for everyone, *regardless of the intent behind these violations*".